requirements.txt
asgiref==3.6.0
Django==3.2.9
django-filter==22.1
djangorestframework==3.12.4
psycopg2-binary==2.9.1
pytz==2023.3
sqlparse==0.4.4
